refactor(GH17): component folders, TodoItem type extraction, branded icons
- Move each component into its own folder with co-located tests, stories,
  and an index.ts barrel — establishes scalable pattern for template consumers
- Extract TodoItem type from todo.worker.ts into src/types/todo.ts; update
  all imports across components, hooks, and worker
- Update manifest.webmanifest to reference todo-pwa-192x192/512x512 icons
- Add new branded PNG icons (todo-pwa-192x192.png, todo-pwa-512x512.png)
- Update ARCHITECTURE.md and STACK.md to document folder-per-component
  convention and template intent
