+++ apps/todo-pwa/public/manifest.webmanifest
@@ -8,17 +8,17 @@
   "theme_color": "#6366f1",
   "icons": [
     {
-      "src": "/icons/pwa-192x192.png",
+      "src": "/icons/todo-pwa-192x192.png",
       "sizes": "192x192",
       "type": "image/png"
     },
     {
-      "src": "/icons/pwa-512x512.png",
+      "src": "/icons/todo-pwa-512x512.png",
       "sizes": "512x512",
       "type": "image/png"
     },
     {
-      "src": "/icons/pwa-512x512.png",
+      "src": "/icons/todo-pwa-512x512.png",
       "sizes": "512x512",
       "type": "image/png",
       "purpose": "maskable"
