+++ apps/todo-pwa/src/components/TodoApp/index.ts
@@ -1,0 +1,1 @@
+export { TodoApp } from "./TodoApp";
