+++ apps/todo-pwa/src/components/TodoInput/index.ts
@@ -1,0 +1,1 @@
+export { TodoInput } from "./TodoInput";
