+++ apps/todo-pwa/src/components/TodoItem/TodoItem.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, it, expect, vi } from "vitest";
 import { TodoItem } from "./TodoItem";
-import type { TodoItem as TodoItemType } from "../workers/todo.worker";
+import type { TodoItem as TodoItemType } from "../../types/todo";
 
 const makeTodo = (overrides: Partial<TodoItemType> = {}): TodoItemType => ({
   id: "1",
