+++ apps/todo-pwa/src/types/todo.ts
@@ -1,0 +1,6 @@
+export type TodoItem = {
+  id: string;
+  description: string;
+  completed: boolean;
+  createdAt: string;
+};
